sleeping restoration of health, magicka, and fatigue

+++ components/esm/loadgmst.cpp
@@ -186,7 +186,7 @@ int GameSetting::getInt() const
     }
 }
 
-int GameSetting::getFloat() const
+float GameSetting::getFloat() const
 {
     switch (type)
     {
